+++ templater/lang/pt-br/lang.php
@@ -1,11 +1,12 @@
 <?php
 /**
- * Brazilian portuguese language file for template plugin
+ * Brazilian portuguese language file for templater plugin
  *
  * @author Daniel Dias Rodrigues <danieldiasr@gmail.com>
  */
-$lang['template_not_found'] = 'O template "%1$s" não foi encontrado. ';
-$lang['click_create'] = '[Clique aqui para criá-lo]';
-$lang['click_edit'] = '[Clique aqui para editá-lo]';
-$lang['stopped_recursion'] = 'O processamento do template "%1$s" foi interrompido devido à recursão. ';
-$lang['no_permissions_view'] = '<div class="template"> Sem permissões para visualizar o template </div>';
+$lang['template'] = '— O template ';
+$lang['not_found'] = ' não foi encontrado.';
+$lang['processing_template'] = '— O processamento do template ';
+$lang['stopped_recursion'] = ' foi interrompido devido à recursão.';
+$lang['no_permissions_view'] = '<div class="templater"> Sem permissões para visualizar o template </div>';
+$lang['no_such_section'] = '<div class="templater">— Não há nenhuma seção "%1$s" na página ';